sound sources correctness

+++ src/sound/player.cpp
@@ -271,7 +271,7 @@ void Player::handleProcessedSamples()
 	alGetSourcei(_private->source, AL_SAMPLE_OFFSET, &sampleOffset);
 	checkOpenALError("alGetSourcei(%d, AL_SAMPLE_OFFSET, %d)", _private->source, sampleOffset);
 	
-	if ((sampleOffset == track()->samples()) && !_private->playingLooped)
+	if ((sampleOffset == static_cast<int>(track()->samples())) && !_private->playingLooped)
 	{
 		stop();
 		finished.invokeInMainRunLoop(this);
